Expose OS Doppler loader from dev module
